pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "woodpecker"
version = "0.0.0"
description = "Lightweight fix catalog + scaffolding for climate dataset fixes"
requires-python = ">=3.10"
dependencies = [
  "click>=8.1",
  "lxml>=4.9",
  "pydantic>=2.0",
  "prov>=2.0",
  "PyYAML>=6.0",
  "rdflib>=7.0",
  "xarray>=2024.1.0",
]

[project.optional-dependencies]
io = [
  "netCDF4>=1.6",
  "h5netcdf>=1.3",
  "scipy>=1.11",
]
zarr = [
  "zarr>=2.13.0,<3.0",
  "numcodecs>=0.12",
  "fsspec>=2024.1.0",
]
docs = [
  "mkdocs<2",
  "mkdocs-material>=9",
  "jinja2>=3.1",
]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
]

[project.scripts]
woodpecker = "woodpecker.cli:cli"

[tool.setuptools.packages.find]
include = ["woodpecker*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "io_backend: tests that require optional NetCDF/Zarr backends",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
